drop Python 3.7 and add 3.11 (#1842)
* drop 3.7 and add 3.11

* Update github_test_action.yml

update to the latest versions of the setuppython and checkout actions

* Update github_test_action.yml

Install numba only for Python Versions < 3.11

* remove deprecated np.float, np.int

use built-in float, int instead

* changelog

* increase timeout for jupyter tests to 15 min per cell (due to pandamodels tests requiring more time to run sometimes)

---------

+++ pandapower/grid_equivalents/rei_generation.py
@@ -289,7 +289,7 @@ def _create_net_zpbn(net, boundary_buses, all_internal_buses, all_external_buses
                 value_types = net[elm][c][elm_org.bus.isin(all_external_buses)].apply(type).unique()
                 if len(value_types) > 1:
                     other_cols_mixed |= {c}
-                elif value_types[0] in (np.float, np.int):
+                elif value_types[0] in (float, int):
                     other_cols_number |= {c}
                 elif value_types[0] == bool:
                     other_cols_bool |= {c}
